fix: trailing slash for dir entries
Previously, we were cleaning the name from the header, and writing that
out. For example, if a TAR entry for "./" was read, it would be written
as ".". Unfortunately, a "." TAR entry causes sqfstar to fail when it is
used as the converter for SquashfsLayer.

Fix this by always adding a trailing slash to directory entries, which
is the norm for TAR streams.