mutate: stream tar output from sqfs2tar
Stream the tar output from sqfs2tar when performing a SquashFS to TAR
layer conversion. This removes the need for one temporary file.

Although intended, we cannot remove the SquashFS temporary file by using
a named pipe for intput to sqfs2tar. Due to the structure of SquashFS
filesystems, sqfs2tar re-reads portions of the input and cannot work
from a named pipe that is streaming input.

Temporary directory is also now passed by option, per comments in review
of #52.

Closes #57