feat: expose lower-level OCIFileImage interface
To permit interaction with an OCI-SIF at a lower-level than the
`v1.ImageIndex` representation of the `RootIndex` blob, expose an
`OCIFileImage` that is analogous to the `layout.Path` in
google/go-containerregistry.

This PR does not add new functionality - it just exposes existing
functionality in a manner that matches the ggcr `layout.Path` interface
as closely as possible.

The existing private `fileImage` struct is made public, and renamed to
`OCIFileImage` so that it is less confusing in consumer code that
generally also uses `sif.FileImage`.

The `sif.FileImage` in `OCIFileImage` is explicitly made a named private
field, so that we do not expose `sif.FileImage` functionality from
`OCIFileImage`.

The private `imageIndex` struct is renamed to `rootIndex` to make it
more clear it refers to the RootIndex descriptor in a SIF, and not any
other ImageIndex in the SIF.

A `FromFileImage` function is added to construct an `OCIFileImage` from
a `sif.FileImage`.

The private `fileImage.writeBlobToFileImage`, `fileImage.writeImageToFileImage`,
and `fileImage.writeIndexToFileImage` methods become public methods of
`OCIFileImage` now named `WriteBlob`, `WriteImage`and `WriteIndex`. These
match ggcr `layout.Path` behaviour - i.e. they don't update the `rootIndex`.
We will add Append / Remove functions which do so at a later PR.

The implementation of `Update` is moved to `OCIFileImage.Update`.

For backward compatibility, a deprecated `Update` function with the
prior signature is provided.

Fixes #80