refactor: improve test corpus
Use AppendImage/AppendIndex in gen_images.go. This involves a meaningful
change in the hello-world-docker-v2-manifest-list layout, which now has
a root index.json that points to an index manifest. Previously, the
hello-world manifest list was used as the root index, which resulted in
an invalid index.json, since it does not use an OCI media type.

The change to AppendImage/AppendIndex has resulted in index.json files
with formatting that is more friendly. In addition to the corpus files
changing, several .golden files have also chaged in tandem.