feat: pkg/ociplatform for OCI platform comparison
OCI platform handling in ggcr is sometimes incomplete / naive. For
example, handling of the variant is not consistent with the
normalisation rules that are applied in e.g. Docker, containerd etc.

Add functions in a new `pkg/ociplatform` package that perform platform
comparison using `github.com/containerd/platforms`, with translation
through ggcr <-> spec-go types as necessary.